# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "USC App Degree Projects API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List Projects",
        "operationId": "list_projects_api_v1_projects_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Status"
            }
          },
          {
            "name": "modality_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Modality Id"
            }
          },
          {
            "name": "academic_period",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Academic Period"
            }
          },
          {
            "name": "academic_program_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Academic Program Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedProjectsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Create Project",
        "description": "Inscribe una nueva idea de trabajo de grado (solo Estudiante).",
        "operationId": "create_project_api_v1_projects_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{project_id}/library-authorization": {
      "patch": {
        "tags": [
          "act"
        ],
        "summary": "Set Library Authorization",
        "description": "El estudiante autoriza (o revoca) la publicación en biblioteca.\nSolo disponible en estado trabajo_aprobado.\nEs idempotente: crea o actualiza el registro en acts.",
        "operationId": "set_library_authorization_api_v1_projects__project_id__library_authorization_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryAuthorizationUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/act": {
      "post": {
        "tags": [
          "act"
        ],
        "summary": "Issue Act",
        "description": "El Administrador emite el acta de sustentación/aprobación.\nRequiere que el estudiante haya diligenciado la autorización de biblioteca.\nEl archivo PDF es opcional.",
        "operationId": "issue_act_api_v1_projects__project_id__act_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_issue_act_api_v1_projects__project_id__act_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "act"
        ],
        "summary": "Get Act",
        "description": "Retorna el detalle del acta. Si existe act_file_url, genera URL firmada (TTL 1h).",
        "operationId": "get_act_api_v1_projects__project_id__act_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "issued_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issued At"
          },
          "issued_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issued By"
          },
          "library_authorization": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Library Authorization"
          },
          "act_file_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Act File Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "issued_at",
          "issued_by",
          "library_authorization",
          "act_file_url"
        ],
        "title": "ActResponse",
        "example": {
          "act_file_url": "https://storage.supabase.co/object/sign/acts/acta.pdf?token=eyJ...",
          "id": "a0b1c2d3-abcd-1234-5678-9abcdef01234",
          "issued_at": "2026-05-20T10:00:00Z",
          "issued_by": "1a2b3c4d-1234-5678-9abc-def012345678",
          "library_authorization": true,
          "project_id": "f6a7b8c9-6789-abcd-f012-345678901234"
        }
      },
...
      "Body_issue_act_api_v1_projects__project_id__act_post": {
        "properties": {
          "file": {
            "anyOf": [
              {
                "type": "string",
                "format": "binary"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          }
        },
        "type": "object",
        "title": "Body_issue_act_api_v1_projects__project_id__act_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LibraryAuthorizationUpdate": {
        "properties": {
          "library_authorization": {
            "type": "boolean",
            "title": "Library Authorization"
          }
        },
        "type": "object",
        "required": [
          "library_authorization"
        ],
        "title": "LibraryAuthorizationUpdate",
        "example": {
          "library_authorization": true
        }
      },
...
      "PaginatedProjectsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProjectResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size"
        ],
        "title": "PaginatedProjectsResponse",
        "example": {
          "items": [],
          "page": 1,
          "size": 20,
          "total": 0
        }
      },
...
      "ProjectCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 100,
            "title": "Title"
          },
          "modality_id": {
            "type": "string",
            "format": "uuid",
            "title": "Modality Id"
          },
          "academic_program_id": {
            "type": "string",
            "format": "uuid",
            "title": "Academic Program Id"
          },
          "research_group": {
            "type": "string",
            "title": "Research Group"
          },
          "research_line": {
            "type": "string",
            "maxLength": 200,
            "title": "Research Line"
          },
          "suggested_director": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Director"
          },
          "member_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "minItems": 1,
            "title": "Member Ids"
          },
          "prerequisite_declaration": {
            "type": "boolean",
            "title": "Prerequisite Declaration"
          }
        },
        "type": "object",
        "required": [
          "title",
          "modality_id",
          "academic_program_id",
          "research_group",
          "research_line",
          "member_ids",
          "prerequisite_declaration"
        ],
        "title": "ProjectCreate",
        "example": {
          "academic_program_id": "a1b2c3d4-1234-5678-abcd-ef0123456789",
          "member_ids": [
            "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          ],
          "modality_id": "b2c3d4e5-2345-6789-bcde-f01234567890",
          "prerequisite_declaration": true,
          "research_group": "GIEIAM",
          "research_line": "Ingeniería de software y sistemas de información",
          "suggested_director": "Dr. Carlos Rodríguez",
          "title": "Sistema de gestión de trabajos de grado USC"
        }
      },
...
      "ProjectResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "modality_id": {
            "type": "string",
            "format": "uuid",
            "title": "Modality Id"
          },
          "academic_program_id": {
            "type": "string",
            "format": "uuid",
            "title": "Academic Program Id"
          },
          "research_group": {
            "type": "string",
            "title": "Research Group"
          },
          "research_line": {
            "type": "string",
            "title": "Research Line"
          },
          "suggested_director": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Director"
          },
          "period": {
            "type": "string",
            "title": "Period"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "has_company_link": {
            "type": "boolean",
            "title": "Has Company Link"
          },
          "plagiarism_suspended": {
            "type": "boolean",
            "title": "Plagiarism Suspended"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "modality_id",
          "academic_program_id",
          "research_group",
          "research_line",
          "suggested_director",
          "period",
          "status",
          "has_company_link",
          "plagiarism_suspended",
          "created_at",
          "updated_at"
        ],
        "title": "ProjectResponse",
        "example": {
          "academic_program_id": "a1b2c3d4-1234-5678-abcd-ef0123456789",
          "created_at": "2026-02-05T09:00:00Z",
          "has_company_link": false,
          "id": "f6a7b8c9-6789-abcd-f012-345678901234",
          "modality_id": "b2c3d4e5-2345-6789-bcde-f01234567890",
          "period": "2026-1",
          "plagiarism_suspended": false,
          "research_group": "GIEIAM",
          "research_line": "Ingeniería de software y sistemas de información",
          "status": "idea_inscrita",
          "suggested_director": "Dr. Carlos Rodríguez",
          "title": "Sistema de gestión de trabajos de grado USC",
          "updated_at": "2026-02-05T09:00:00Z"
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}